requirements.txt
bs4
discord
gspread
lxml
psycopg2
python-dotenv
PyYAML
requests